// frontend/src/client/types.gen.ts
export type ReviewCreate = {
    /**
     * Rating
     */
    rating: number;
    /**
     * Comment
     */
    comment?: string | null;
};